requirements.txt
ultralytics>=8.3.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
pyyaml>=6.0.0
tqdm>=4.66.0
scikit-learn>=1.3.0
gradio>=4.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
